Testcases typescript (#124)
* initial commit

* converted testcases to ts

* Update test/mainnet/aave/v1.test.ts


* Update test/mainnet/aave/v1.test.ts


* Update test/mainnet/aave/v1.test.ts


* updated test

* updated-test

* removed typechain
