+++ test/mainnet/aave/v1.test.ts
@@ -1,28 +1,29 @@
-const { expect } = require("chai");
-const hre = require("hardhat");
-const abis = require("../../../scripts/constant/abis");
-const addresses = require("../../../scripts/constant/addresses");
-const deployAndEnableConnector = require("../../../scripts/deployAndEnableConnector");
-const getMasterSigner = require("../../../scripts/getMasterSigner");
-const buildDSAv2 = require("../../../scripts/buildDSAv2");
-const ConnectV2AaveV1 = require("../../artifacts/contracts/mainnet/connectors/aave/v1/main.sol/ConnectV2AaveV1.json");
-const { parseEther } = require("@ethersproject/units");
-const encodeSpells = require("../../../scripts/encodeSpells");
-const tokens = require("../../../scripts/constant/tokens");
-const constants = require("../../../scripts/constant/constant");
-const addLiquidity = require("../../../scripts/addLiquidity");
-const { ethers } = hre;
+import { expect } from "chai";
+import "hardhat";
+import { abis } from "../../../scripts/constant/abis";
+import { addresses } from "../../../scripts/constant/addresses";
+import { deployAndEnableConnector } from "../../../scripts/deployAndEnableConnector";
+import { getMasterSigner } from "../../../scripts/getMasterSigner";
+import { buildDSAv2 } from "../../../scripts/buildDSAv2";
+
+import ConnectV2AaveV1 from "../../artifacts/contracts/mainnet/connectors/aave/v1/main.sol/ConnectV2AaveV1.json";
+import { parseEther } from "@ethersproject/units";
+import { encodeSpells } from "../../../scripts/encodeSpells";
+import { tokens } from "../../../scripts/constant/tokens";
+import { constants } from "../../../scripts/constant/constant";
+import { addLiquidity } from "../../../scripts/addLiquidity";
+// const { ethers } = hre;
 
 const ALCHEMY_ID = process.env.ALCHEMY_ID;
 
-describe("Aave V1", function() {
+describe("Aave V1", function () {
   const connectorName = "AAVEV1-TEST-A";
 
-  let wallet0, wallet1;
-  let dsaWallet0;
-  let instaConnectorsV2;
-  let connector;
-  let masterSigner;
+  let wallet0: any, wallet1: any;
+  let dsaWallet0: any;
+  let instaConnectorsV2: any;
+  let connector: any;
+  let masterSigner: any;
 
   before(async () => {
     try {
@@ -61,13 +62,13 @@ describe("Aave V1", function() {
     expect(!!masterSigner.address).to.be.true;
   });
 
-  describe("DSA wallet setup", function() {
-    it("Should build DSA v2", async function() {
+  describe("DSA wallet setup", function () {
+    it("Should build DSA v2", async function () {
       dsaWallet0 = await buildDSAv2(wallet0.address);
       expect(!!dsaWallet0.address).to.be.true;
     });
 
-    it("Deposit ETH into DSA wallet", async function() {
+    it("Deposit ETH into DSA wallet", async function () {
       await wallet0.sendTransaction({
         to: dsaWallet0.address,
         value: parseEther("10"),
@@ -78,8 +79,8 @@ describe("Aave V1", function() {
     });
   });
 
-  describe("Main", function() {
-    it("should deposit ETH in Aave V1", async function() {
+  describe("Main", function () {
+    it("should deposit ETH in Aave V1", async function () {
       const amt = parseEther("1");
       const spells = [
         {
@@ -100,7 +101,7 @@ describe("Aave V1", function() {
       );
     });
 
-    it("Should borrow and payback DAI from Aave V1", async function() {
+    it("Should borrow and payback DAI from Aave V1", async function () {
       const amt = parseEther("100"); // 100 DAI
 
       // add a little amount of dai to cover any shortfalls
@@ -129,7 +130,7 @@ describe("Aave V1", function() {
       );
     });
 
-    it("Should deposit all ETH in Aave V1", async function() {
+    it("Should deposit all ETH in Aave V1", async function () {
       const spells = [
         {
           connector: connectorName,
@@ -147,7 +148,7 @@ describe("Aave V1", function() {
       );
     });
 
-    it("Should withdraw all ETH from Aave V1", async function() {
+    it("Should withdraw all ETH from Aave V1", async function () {
       const spells = [
         {
           connector: connectorName,
